feat: merge http registry into logos delivery
