+++ crates/generic-chat/src/lib.rs
@@ -24,6 +24,5 @@ pub use logos_account::AccountDirectory;
 // Re-export bundled registry implementations so callers can pick one without
 // pulling in `components` directly.
 pub use components::{
-    DeliveryRegistry, DeliveryRegistryError, EphemeralRegistry, HttpRegistry, HttpRegistryError,
-    RegistryPublishMode,
+    DeliveryRegistry, DeliveryRegistryError, EphemeralRegistry, RegistryPublishMode,
 };
