+++ extensions/components/src/contact_registry.rs
@@ -1,3 +1,2 @@
 pub mod delivery;
 pub mod ephemeral;
-pub mod http;
