+++ extensions/components/src/lib.rs
@@ -8,7 +8,6 @@ pub use contact_registry::delivery::{
     RegistryPublishMode,
 };
 pub use contact_registry::ephemeral::EphemeralRegistry;
-pub use contact_registry::http::{HttpRegistry, HttpRegistryError};
 pub use delivery::*;
 pub use storage::*;
 pub use wakeup::*;
